requirements.txt
discord.py
selenium
datetime